Renamed to add _dir; Extracted find_build_dir func
