--- conf/script/src/build_system/cmd/clean_build/__init__.py
@@ -1,2 +1,0 @@
-import build_system.cmd.clean_build.cli
-from build_system.cmd.clean_build.clean_build import clean_build
