+++ conf/script/src/build_system/cmd/clean_build_dir/__init__.py
@@ -1,0 +1,2 @@
+import build_system.cmd.clean_build_dir.cli
+from build_system.cmd.clean_build_dir.clean_build_dir import clean_build_dir
